requirements.txt
customtkinter
pillow
ultralytics
opencv-python-headless
numpy